step-style: add <a><paper-button> styling
All this time we've had the following "download" paper-button structure
on the official website at g.co/codelabs:

    <a href="..."><paper-button>download</paper-button></a>

Note that this is in accordance with the Polymer 1.0 recommendation
in https://elements.polymer-project.org/bower_components/paper-button/demo/index.html

Unfortunately, this is not how codelab-components were styled all this
time. A demo codelab has the inverse structure:

    <paper-button><a href="...">download</a></paper-button>

Here's where it's used in codelab-components repo:
https://github.com/googlecodelabs/codelab-components/blob/7397f708d210dfbb7348084033a32ba070ec4df4/demo/codelab.html#L74

As a consequence, the styling Will provided in a recent change in
https://github.com/googlecodelabs/codelab-components/commit/f3e63d5612e67e9e35804f0992807c4aec5cbd8e
used the latter structure. We never noticed the difference because
before the styling refresh, the download button background and link
colors were very close.